feat: 将`打开配置目录`按钮移到下载目录旁边
