+++ src/AppContent.vue
@@ -6,8 +6,6 @@ import QrcodeViewer from "./components/QrcodeViewer.vue";
 import DownloadingList from "./components/DownloadingList.vue";
 import SearchPane from "./components/SearchPane.vue";
 import EpisodePane from "./components/EpisodePane.vue";
-import {appDataDir} from "@tauri-apps/api/path";
-import {path} from "@tauri-apps/api";
 import CookieLoginDialog from "./components/CookieLoginDialog.vue";
 
 const notification = useNotification();
@@ -56,15 +54,6 @@ onMounted(async () => {
   config.value = await commands.getConfig();
 });
 
-async function showConfigInFileManager() {
-  const configName = "config.json";
-  const configPath = await path.join(await appDataDir(), configName);
-  const result = await commands.showPathInFileManager(configPath);
-  if (result.status === "error") {
-    notification.error({title: "打开配置文件失败", description: result.error});
-  }
-}
-
 async function test() {
   notification.error({
     title: "标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 标题 ",
@@ -93,7 +82,6 @@ async function test() {
         <div class="flex">
           <n-button @click="qrcodeViewerShowing=true" type="primary">二维码登录</n-button>
           <n-button @click="cookieLoginDialogShowing=true" type="primary" secondary>Cookie登录</n-button>
-          <n-button @click="showConfigInFileManager">打开配置目录</n-button>
           <n-button @click="test">测试用</n-button>
           <div v-if="userProfile!==undefined" class="flex flex-justify-end">
             <n-avatar round
